requirements.txt
flask==2.3.3
python-dotenv==1.0.0
pymongo==4.5.0
chromadb==0.4.3
python-jose[cryptography]==3.3.0
passlib==1.7.4
flask-cors==4.0.0
requests==2.31.0 